Struct libcw::simulation::Mars

pub struct Mars { /* fields omitted */ }

A corewars simulator

Methods

impl Mars

pub fn pc(&self) -> Option<Address>

Return the next program counter

pub fn size(&self) -> usize

Return size of mars' memory

pub fn cycle(&self) -> usize

Return the currect cpu cycle

pub fn pid(&self) -> Option<Pid>

Return the next process id that will execute

pub fn pids(
&self
) -> impl Iterator<Item = Pid> + '_

Return all active procces ids

pub fn processes(
&self
) -> impl Iterator<Item = &Process>

Returns each processes resources zipped with its pid

pub fn process_queues(
&self
) -> impl Iterator<Item = (Pid, impl Iterator<Item = &Address>)>

Return process queues zipped with the owning process' id

pub fn process_pspaces(
&self
) -> impl Iterator<Item = (usize, &[Address])>

Return process private storage(pspace) zipped with the owning process' id

pub fn process_count(&self) -> usize

Returns the current number of processes

pub fn thread_count(
&self
) -> impl Iterator<Item = (Pid, usize)> + '_

Return the number of threads each process has along with the process id

pub fn memory(&self) -> &[Instruction]

Return a view of the Mars' memory

pub fn set_memory(&mut self, instructions: &[Instruction], address: Address)

pub fn load_program(&mut self, program: &[Instruction], address: Address) -> Pid

pub fn load_programs_with_shared_pspace(
&mut self,
programs: &[&[Instruction]],
addresses: &[Address]
) -> Vec<Pid>

Load multiple programs in different locations with the same pspace

Returns

A slice containing all of the created process ids, in order

pub fn step(&mut self) -> Option<Pid>

Step forward one clock cycle

Panics

panics if there are no processes in the Mars

Returns

Some(pid) if a process with id pid was killed. Otherwise None

pub fn reset(&mut self) -> &mut Self

Reset the mars

Clears the processe queue and memory