Struct libcw::simulation::MarsBuilder

pub struct MarsBuilder { /* fields omitted */ }

A Mars builder. Provides control over how the Mars is configured

Methods

impl MarsBuilder

fn new() -> Self

Create a MarsBuilder with default parameters

fn build_and_load<T>(
&self,
programs: Vec<(Address, Option<Pin>, &Vec<T>)>
) -> LoadResult<Mars<T>> where
T: Instruction,

Build a core and load it with specified programs

fn build<T>(&self) -> Mars<T> where
T: Instruction,

Build a halted mars

fn size(&mut self, size: usize) -> &mut Self

Size of memory

Arguments

size: size of memory

Return

Self

fn pspace_size(&mut self, size: usize) -> &mut Self

Size of each warrior's P-space

Arguments

size: size of memory

Return

Self

fn max_cycles(&mut self, n: usize) -> &mut Self

Maximum number of cycles that can elapse before a tie is declared

Arguments

n: number of cycles

Return

Self

fn max_processes(&mut self, n: usize) -> &mut Self

Maximum number of processes a core can have in it's process queue

Arguments

n: number of processes

Return

Self

fn max_length(&mut self, n: usize) -> &mut Self

Maximum number of instructions allowed in a program

Arguments

n: number of instructions

Return

Self

fn min_distance(&mut self, n: usize) -> &mut Self

Minimum distance between warriors

Arguments

n: number of instructions

Return

Self

fn version(&mut self, version: usize) -> &mut Self

Mars version multiplied by 100 (e.g. version 0.8 -> 80)

Arguments

version: version number

Return

Self