Struct libcw::simulation::Mars

pub struct Mars<T> where
    T: Instruction,  { /* fields omitted */ }

Core wars runtime

Methods

impl<T> Mars<T> where
T: Instruction,

fn step(&mut self) -> SimulationResult<SimulationEvent>

Step forward one cycle

fn halted(&self) -> bool

Has the core finished its execution. This can mean either a tie has occurred or a warrior has emerged victoriors

fn reset(&mut self)

Reset the Mars's memory and the process queue

fn reset_hard(&mut self)

Reset the Mar's memory, process queue, AND P-space

fn load_batch(
&mut self,
programs: Vec<(Address, Option<Pin>, &Vec<T>)>
) -> LoadResult<()>

Load mutliple programs into the Mars, checking their spacing and their length

Arguments

programs: programs and load information loaded in a tuple, cannot be empty # Return Ok(()) if the load was successful, otherwise an error with the corresponding SimulationError

fn pc(&self) -> Address

Get Pid currently executing on the core

Panics

Panics is the process queue is empty

fn pcs(&self) -> Vec<Address>

Get the program counters for all processes

fn cycle(&self) -> usize

Current cycle core is executing

fn pid(&self) -> Pid

Get the current Pid executing

fn pids(&self) -> Vec<Pid>

Get all Pids that are currently active in the order they will be executing

fn size(&self) -> usize

Size of memory

fn pspace_size(&self) -> usize

Size of private storage space

fn version(&self) -> usize

Version of core multiplied by 100

fn max_processes(&self) -> usize

Maximum number of processes that can be in the core queue

fn max_cycles(&self) -> usize

Maximum number of cycles before a tie is declared

fn max_length(&self) -> usize

Maximum number of instructions allowed in a program

fn min_distance(&self) -> usize

Minimum distance allowed between programs

fn memory(&self) -> &Vec<T>

Get immutable reference to memory

fn pspace(&self) -> &HashMap<Pin, Vec<Value>>

Get an immutable reference to private storage

fn process_count(&self) -> usize

Get the number of processes currently running

fn current_queue(&self) -> Option<&VecDeque<Address>>

Fetch reference to current queue